Add files via upload
The sample projects that I have supervised and guided as a mentor to bachelor and master students in the Data Analysis Industry Training Program. We will be discussing the fundamental stages of data analysis over the course of 4 weeks. I would like to express my admiration for the diligent and innovative reports submitted by the bright students in our class, who hail from esteemed international universities and have dedicated 5 hours each week to completing the program and their reports.